+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/request/internal/RequestExecutor.kt
@@ -16,9 +16,9 @@
 
 package com.github.panpf.sketch.request.internal
 
+import androidx.annotation.MainThread
 import com.github.panpf.sketch.Image
 import com.github.panpf.sketch.Sketch
-import com.github.panpf.sketch.annotation.MainThread
 import com.github.panpf.sketch.request.DepthException
 import com.github.panpf.sketch.request.ImageData
 import com.github.panpf.sketch.request.ImageRequest