+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/request/internal/RequestInterceptorChain.kt
@@ -16,7 +16,7 @@
 
 package com.github.panpf.sketch.request.internal
 
-import com.github.panpf.sketch.annotation.MainThread
+import androidx.annotation.MainThread
 import com.github.panpf.sketch.request.ImageData
 import com.github.panpf.sketch.request.ImageRequest
 import com.github.panpf.sketch.request.RequestContext