+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/request/internal/RequestManager.kt
@@ -16,8 +16,8 @@
 
 package com.github.panpf.sketch.request.internal
 
+import androidx.annotation.MainThread
 import com.github.panpf.sketch.Sketch
-import com.github.panpf.sketch.annotation.MainThread
 import com.github.panpf.sketch.request.Disposable
 import com.github.panpf.sketch.request.ImageRequest
 import com.github.panpf.sketch.request.ImageResult