+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/transition/Transition.kt
@@ -17,8 +17,8 @@
 
 package com.github.panpf.sketch.transition
 
+import androidx.annotation.MainThread
 import com.github.panpf.sketch.Sketch
-import com.github.panpf.sketch.annotation.MainThread
 import com.github.panpf.sketch.request.ImageRequest
 import com.github.panpf.sketch.request.ImageResult
 import com.github.panpf.sketch.target.Target